Revert "Merge remote-tracking branch 'origin' into features/topology"
This reverts commit 3b3c99c1547bbd705efc4a74c612dbbc55de3564, reversing
changes made to 56974677f65d0eaaba5bdcf2e62e7acfde204e27.

+++ features/enlinkd/adapters/updaters/cdp/src/main/java/org/opennms/netmgt/enlinkd/CdpOnmsTopologyUpdater.java
@@ -95,7 +95,7 @@ public class CdpOnmsTopologyUpdater extends TopologyUpdater {
         for(TopologyConnection<CdpLinkTopologyEntity, CdpLinkTopologyEntity> pair : m_cdpTopologyService.match()) {
             topology.getEdges().add(
                 OnmsTopologyEdge.create(
-                        Topology.getDefaultEdgeId(pair.getLeft().getId(), pair.getRight().getId()),
+                        Topology.getDefaultEdgeId(pair.getLeft().getId().intValue(), pair.getRight().getId().intValue()),
                         create(topology.getVertex(pair.getLeft().getNodeIdAsString()),
                                pair.getLeft(),
                                nodeToOnmsSnmpTable.get(pair.getLeft().getNodeId(), pair.getLeft().getCdpCacheIfIndex())),
